refactor: use the same prepareTemplateData func
